COMP: Fix compiler warnings from Qt 6 and VTK 9 deprecated APIs
Address deprecation warnings reported by clang when building against
Qt 6.10 and current VTK, while remaining compatible with Qt 5.15:

- Replace QVariant::type()/QVariant::Type comparisons with userType()
  and QMetaType enums, and deprecated QVariant::canConvert(int) with
  the templated canConvert<T>().
- Replace deprecated QScopedPointer::take() with std::unique_ptr and
  release().
- Replace deprecated QKeySequence modifier operator+ with operator|.
- Guard Qt::AA_EnableHighDpiScaling/AA_UseHighDpiPixmaps (no-op and
  deprecated in Qt 6) and use QImage::flipped() on Qt >= 6.9.
- Use the QMouseEvent constructor that takes a global position.
- Replace deprecated vtkCellArray methods: GetCell() with
  GetCellAtId(), Allocate(EstimateSize())/Allocate(GetSize()) with
  AllocateEstimate()/AllocateExact(), ReplaceCell() with
  ReplaceCellAtId(), and remove no-op SetNumberOfCells() and
  vestigial traversal calls.
- Replace VTK_FALLTHROUGH with [[fallthrough]].
- Replace sprintf() with snprintf() (deprecated on macOS for security
  reasons).
- Replace deprecated CURLOPT_PUT with CURLOPT_UPLOAD.
- Check QFile::open() return values (nodiscard warnings).
- Add missing override to vtkMRMLTransformSequenceStorageNode
  ::SupportedFileType() and make int64-to-double conversion explicit
  in vtkITKArchetypeImageSeriesReader.

+++ Libs/vtkITK/vtkITKArchetypeImageSeriesReader.cxx
@@ -903,8 +903,8 @@ int vtkITKArchetypeImageSeriesReader::RequestInformation(vtkInformation* vtkNotU
           {
             scalarType = VTK_INT;
           }
-          else if (max <= std::numeric_limits<int64_t>::max() //
-                   && min >= std::numeric_limits<int64_t>::min())
+          else if (max <= static_cast<double>(std::numeric_limits<int64_t>::max()) //
+                   && min >= static_cast<double>(std::numeric_limits<int64_t>::min()))
           {
             scalarType = VTK_LONG;
           }
